+++ src/_scss/pages/search/resultsView/_resultsView.scss
@@ -131,6 +131,28 @@
         margin-bottom: rem(16);
       }
 
+      .download {
+        display: flex;
+        justify-content: end;
+        margin-bottom: rem(8);
+        font-size: rem(12);
+        line-height: 1.5;
+        color: #2378c3;
+
+        span.text {
+          margin-left: 4px;
+        }
+
+        .tooltip__hover-wrapper {
+          display: flex;
+
+          svg {
+            width: 18px;
+            height: 18px;
+          }
+        }
+      }
+
       .visualization-pager-container {
         margin-top: rem(16);
       }
