+++ src/js/containers/search/newResultsView/TimeVisualizationSectionContainer.jsx
@@ -7,10 +7,6 @@ import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { isCancel } from 'axios';
-import { TooltipWrapper } from "data-transparency-ui";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { capitalize } from "lodash";
-import { fullMonthFromAbbr } from 'helpers/monthHelper';
 
 import * as searchFilterActions from 'redux/actions/search/searchFilterActions';
 import { setAppliedFilterCompletion } from 'redux/actions/search/appliedFilterActions';
@@ -19,10 +15,10 @@ import * as SearchHelper from 'helpers/searchHelper';
 import * as MonthHelper from 'helpers/monthHelper';
 
 import SearchAwardsOperation from 'models/v1/search/SearchAwardsOperation';
-import TimeVisualizationChart from "../../../components/search/visualizations/time/TimeVisualizationChart";
 import SearchSectionWrapper from "../../../components/search/newResultsView/SearchSectionWrapper";
 import BaseSpendingOverTimeRow from "../../../models/v2/search/visualizations/time/BaseSpendingOverTimeRow";
 import * as MoneyFormatter from "../../../helpers/moneyFormatter";
+import TimeSectionWrapper from "../../../components/search/newResultsView/time/TimeSectionWrapper";
 
 const combinedActions = Object.assign({}, searchFilterActions, {
     setAppliedFilterCompletion
@@ -249,53 +245,6 @@ const TimeVisualizationSectionContainer = (props) => {
         fetchAwards('Spending Over Time Visualization');
     };
 
-    const getDownloadData = () => {
-        const headers = [];
-        headers.fiscal_year = 'fiscal_year,total_obligations\n';
-        headers.quarter = 'fiscal_year,fiscal_quarter,total_obligations\n';
-        headers.month = 'fiscal_year,month,total_obligations\n';
-        const data = this.props.data;
-
-        return headers[visualizationPeriod].concat(
-            data.rawLabels.map((label, i) => {
-                if (visualizationPeriod === 'fiscal_year') {
-                    return `${label.year},${data.ySeries[i][0]}`;
-                }
-                if (!label.period) { // API still updating data
-                    return null;
-                }
-                if (visualizationPeriod === 'quarter') {
-                    return `${label.year},${label.period[1]},${data.ySeries[i][0]}`;
-                }
-                const month = fullMonthFromAbbr(label.period);
-                return `${['Oct', 'Nov', 'Dec'].indexOf(label.period) > -1 ? parseInt(label.year, 10) + 1 : label.year},${month},${data.ySeries[i][0]}`;
-            })
-                .join('\n')
-        );
-    };
-
-    const downloadBlob = () => new Blob([getDownloadData()], { type: 'text/csv;charset=utf-8;' });
-
-    const renderDownloadLink = () => (
-        <a
-            href={URL.createObjectURL(this.downloadBlob())}
-            download="spending-over-time.csv" >
-            <FontAwesomeIcon icon="download" size="lg" />
-            <span className="text">
-                Download data by {capitalize(visualizationPeriod === 'fiscal_year' ? 'year' : visualizationPeriod)}
-            </span>
-        </a>
-    );
-
-    const downloadTooltip = () => (
-        <>
-            <div className="tooltip__title">Download data by {capitalize(visualizationPeriod === 'fiscal_year' ? 'year' : visualizationPeriod)}</div>
-            <div className="tooltip__text">
-                Download a CSV of award spending data that matches your search criteria, broken down by {visualizationPeriod === 'fiscal_year' ? 'year' : visualizationPeriod}. For complete download results, click on the &quot;Download&quot; button on the top right of this page.
-            </div>
-        </>
-    );
-
     useEffect(() => {
         sortBy("aggregated_amount", "desc");
         // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -343,12 +292,8 @@ const TimeVisualizationSectionContainer = (props) => {
             isError={parsedData?.error}
             hasNoData={parsedData?.ySeries?.flat()?.reduce((partialSum, a) => partialSum + a, 0) === 0}
             manualSort>
-            <div className="download">
-                {!parsedData?.loading && renderDownloadLink()}
-                {!parsedData?.loading && <TooltipWrapper className="tooltip-wrapper" icon="info" tooltipPosition="left" tooltipComponent={this.downloadTooltip()} />}
-            </div>
-            <TimeVisualizationChart
-                {...parsedData}
+            <TimeSectionWrapper
+                parsedData={parsedData}
                 visualizationPeriod={visualizationPeriod}
                 subaward={props.subaward} />
         </SearchSectionWrapper>
